requirements.txt
pillow==10.3.0
python-ev3dev2==2.1.0.post1
